完成JwtUtil和JwtFilter

+++ SpringSecurityDemo/src/main/java/org/bigboss/springsecuritydemo/domain/Member.java
@@ -1,13 +1,11 @@
 package org.bigboss.springsecuritydemo.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.*;
 
 import java.io.Serializable;
+import java.util.Set;
 
 /**
  * @author: maifuwa
@@ -25,7 +23,7 @@ public class Member extends BaseEntity implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String username;
